requirements.txt
langchain==0.0.173
pinecone_client==2.2.1
PyPDF2==3.0.1
python-dotenv==1.0.0
python-telegram-bot==20.3